# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Puzzle API",
    "description": "REST contract for room metadata and WebRTC ICE configuration.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
  "paths": {
    "/api/rooms": {
      "post": {
        "operationId": "createRoom",
        "tags": [
          "Rooms"
        ],
        "summary": "Create a puzzle room",
        "responses": {
          "201": {
            "description": "Created room",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "room": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "difficulty": {
                          "enum": [
                            48,
                            96,
                            192
                          ],
                          "type": "number"
                        },
                        "expiresAt": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "participantCount": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "id",
                        "difficulty",
                        "expiresAt",
                        "participantCount"
                      ]
                    }
                  },
                  "required": [
                    "room"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "410": {
            "description": "Gone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "difficulty": {
                    "enum": [
                      48,
                      96,
                      192
                    ],
                    "type": "number"
                  }
                },
                "required": [
                  "difficulty"
                ]
              }
            }
          }
        }
      }
    },
    "/api/rooms/{roomId}": {
      "get": {
        "operationId": "getRoom",
        "tags": [
          "Rooms"
        ],
        "summary": "Get room metadata",
        "responses": {
          "200": {
            "description": "Room metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "room": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "difficulty": {
                          "enum": [
                            48,
                            96,
                            192
                          ],
                          "type": "number"
                        },
                        "expiresAt": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "participantCount": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "id",
                        "difficulty",
                        "expiresAt",
                        "participantCount"
                      ]
                    }
                  },
                  "required": [
                    "room"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "410": {
            "description": "Gone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 10,
              "pattern": "^[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]+$"
            },
            "required": true
          }
        ]
      }
    },
    "/api/ice": {
      "get": {
        "operationId": "getIceConfig",
        "tags": [
          "Realtime"
        ],
        "summary": "Get WebRTC ICE server configuration",
        "responses": {
          "200": {
            "description": "ICE server configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "iceServers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "urls": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            ]
                          },
                          "username": {
                            "type": "string"
                          },
                          "credential": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "urls"
                        ]
                      }
                    }
                  },
                  "required": [
                    "iceServers"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {}
}